requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.0.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-elasticsearch-dsl==6.5.0
django-js-asset==1.2.2
django-keycloak @ git+https://github.com/Peter-Slump/django-keycloak.git@5211fba26ae4216b4344309eb4a67da9745ada05
django-timezone-field==4.2.1
django-tinymce==3.3.0
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
dnspython==2.1.0
ecdsa==0.17.0
elasticsearch==7.15.1
elasticsearch-dsl==7.4.0
gevent==21.8.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
kombu==5.1.0
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2==2.8.4
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.2
python-jose==3.3.0
python-keycloak-client==0.2.3
pytz==2021.1
redis==3.5.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
